Some parts of codes, files and folder were be refactored
Names of files and folder was changed from "State" to "Screen". Some
parts of codes were be updated.
